# storage/api-docs/api-docs.yaml
openapi: 3.0.0
info:
  title: 'Laravel Base Project API'
  description: 'Laravel Base Project API Documentation'
  version: 1.0.0
servers:
  -
    url: 'http://my-default-host.com'
    description: 'API Server'
paths:
...
  /api/v1/todos:
    get:
      tags:
        - 'FocusFlow Todos'
      summary: 'Get all todos'
      operationId: efcabe2a320c2121610cfe7c5970060c
      parameters:
        -
          name: completed
          in: query
          description: 'Filter by completion status'
          required: false
          schema:
            type: boolean
      responses:
        200:
          description: 'List of todos'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Todo'
      security:
        -
          sanctum: []
    post:
      tags:
        - 'FocusFlow Todos'
      summary: 'Create a new todo'
      operationId: 0bef9d64d28d338779319486730dee88
      requestBody:
        required: true
        content:
          application/json:
            schema:
              required:
                - title
              properties:
                title:
                  type: string
                  example: 'New Task'
                description:
                  type: string
                  nullable: true
                category:
                  type: string
                  example: İş
                priority:
                  type: string
                  enum:
                    - Düşük
                    - Orta
                    - Yüksek
                    - Kritik
                  example: Orta
                deadline:
                  type: string
                  format: date-time
                  nullable: true
                order:
                  type: integer
                  example: 0
                subtasks:
                  type: array
                  items:
                    properties:
                      title:
                        type: string
                        example: 'Subtask 1'
                      order:
                        type: integer
                        example: 0
                    type: object
              type: object
      responses:
        201:
          description: 'Todo created'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
      security:
        -
          sanctum: []
  '/api/v1/todos/{id}':
    get:
      tags:
        - 'FocusFlow Todos'
      summary: 'Get a specific todo'
      operationId: 905e163a9a4d2daee09e5cf98d263d85
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo details'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
      security:
        -
          sanctum: []
    put:
      tags:
        - 'FocusFlow Todos'
      summary: 'Update a todo'
      operationId: 60bac12653e45fc8a2f89bf107c119a9
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                title:
                  type: string
                description:
                  type: string
                  nullable: true
                category:
                  type: string
                priority:
                  type: string
                  enum:
                    - Düşük
                    - Orta
                    - Yüksek
                    - Kritik
                deadline:
                  type: string
                  format: date-time
                  nullable: true
                order:
                  type: integer
                subtasks:
                  type: array
                  items:
                    properties:
                      title:
                        type: string
                      completed:
                        type: boolean
                      order:
                        type: integer
                    type: object
              type: object
      responses:
        200:
          description: 'Todo updated'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
      security:
        -
          sanctum: []
    delete:
      tags:
        - 'FocusFlow Todos'
      summary: 'Delete a todo'
      operationId: ecc770bd143004375538946adb23157a
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo deleted'
      security:
        -
          sanctum: []
  '/api/v1/todos/{id}/complete':
    post:
      tags:
        - 'FocusFlow Todos'
      summary: 'Complete a todo'
      operationId: b8ac4b96a9690b487ecaa7521034fc5b
      parameters:
        -
          name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        200:
          description: 'Todo completed'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Todo'
      security:
        -
          sanctum: []
  /api/v1/todos/reorder:
    post:
      tags:
        - 'FocusFlow Todos'
      summary: 'Reorder todos'
      operationId: 0ba7ad9065e7836e693f092722ec7165
      requestBody:
        required: true
        content:
          application/json:
            schema:
              properties:
                todo_ids:
                  type: array
                  items:
                    type: integer
              type: object
      responses:
        200:
          description: 'Todos reordered'
      security:
        -
          sanctum: []
components:
  schemas:
...
    Todo:
      title: 'Todo Model'
      description: 'FocusFlow Todo model'
      required:
        - title
        - user_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        user_id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Buy milk'
        description:
          type: string
          example: 'Go to the grocery store'
          nullable: true
        category:
          type: string
          example: Shopping
        priority:
          type: string
          enum:
            - Düşük
            - Orta
            - Yüksek
            - Kritik
          example: Orta
        completed:
          type: boolean
          example: false
        deadline:
          type: string
          format: date-time
          nullable: true
        order:
          type: integer
          example: 0
        subtasks:
          type: array
          items:
            $ref: '#/components/schemas/TodoSubtask'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
    TodoSubtask:
      title: 'TodoSubtask Model'
      description: 'FocusFlow Todo Subtask model'
      required:
        - title
        - todo_id
      properties:
        id:
          type: integer
          readOnly: true
          example: 1
        todo_id:
          type: integer
          example: 1
        title:
          type: string
          example: 'Check emails'
        completed:
          type: boolean
          example: false
        order:
          type: integer
          example: 0
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      type: object
  securitySchemes:
    sanctum:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 'Laravel Sanctum token authentication. Enter token in format (Bearer <token>)'
...